+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ### 3.6.0b
  * Key shortcuts (accesskey) for left menu (eg. Chrome ALT+<key>)
+ * Added buttons commands (eg. play, pause) for single file in download
 
 ### 3.3.0b
  * Added Turtle mode: toggle custom limit bandwidth