+++ CHANGELOG.md
@@ -3,7 +3,10 @@ CHANGELOG
 
 ### 3.6.0b
  * Key shortcuts (accesskey) for left menu (eg. Chrome ALT+<key>)
- * Added buttons commands (eg. play, pause) for single file in download
+ * Download: added buttons commands (eg. play, pause) for single file
+ * Search: kad (instead Global) for default
+ * Search: added button downloads for single file result
+ * Search: added spinner (heuristic calculation of finish)
 
 ### 3.3.0b
  * Added Turtle mode: toggle custom limit bandwidth